daemon: Make it possible to pass `substituters` and `builders`
Adds the ability to input additional nix arguments through to the nix
commands. For our sanity and so that we don’t have to replicate the
whole nix API, they are passed as json string on the command line.

However, I don’t know how an exhaustive API would look like (nix has a
lot of configuration options and their semantics is pretty weird in
places), so I opted to only add `--substituters` and `--builders`. We
even gain a nicer interface by using an optional list.

I’m not yet positive on how to document the JSON interface more
nicely, maybe we do want to expose all options as command line flags,
but I really don’t know how that would look without leading to a mess.

The changes are rather simple and mostly mechanical.

As a followup, we should add the same flag to `watch` and `shell`.